extract paper bridge helpers
