+++ crates/apps/reticulumd/src/bin/reticulumd/bridge.rs
@@ -6,6 +6,8 @@ use super::bridge_helpers::{
 mod announce;
 #[path = "bridge_link_send.rs"]
 mod link_send;
+#[path = "bridge_paper.rs"]
+mod paper;
 #[path = "bridge_remote_control.rs"]
 mod remote_control;
 use super::inbound_worker::{
@@ -14,7 +16,6 @@ use super::inbound_worker::{
 use lxmf::WireMessage;
 use rand_core::OsRng;
 use reticulum_daemon::lxmf_bridge::build_wire_message_with_options;
-use reticulum_daemon::lxmf_bridge::rmpv_to_json;
 use reticulum_daemon::lxmf_stamps::generate_propagation_stamp;
 use reticulum_daemon::receipt_bridge::{track_receipt_mapping, ReceiptEvent};
 use rns_core::identity::{Identity as CoreIdentity, PrivateIdentity};
@@ -716,85 +717,11 @@ impl OutboundBridge for TransportBridge {
         &self,
         record: &rns_rpc::MessageRecord,
     ) -> Result<Option<PaperEncodeEnvelope>, std::io::Error> {
-        let destination = parse_destination_hash_required(&record.destination)?;
-        let destination_identity = self
-            .peer_crypto
-            .lock()
-            .expect("peer map")
-            .get(&record.destination)
-            .map(|info| info.identity)
-            .or_else(|| {
-                resolve_destination_identity_blocking(
-                    self.transport.clone(),
-                    AddressHash::new(destination),
-                    Duration::from_secs(12),
-                )
-            });
-        let Some(destination_identity) = destination_identity else {
-            return Ok(None);
-        };
-        let payload = build_wire_message_with_options(
-            self.delivery_source_hash,
-            destination,
-            &record.title,
-            &record.content,
-            record.fields.clone(),
-            &self.signer,
-            None,
-            None,
-            None,
-        )
-        .map_err(std::io::Error::other)?;
-        let wire = WireMessage::unpack(payload.as_slice()).map_err(std::io::Error::other)?;
-        let transient_id = hex::encode(wire.message_id());
-        let destination_identity = CoreIdentity::new_from_slices(
-            destination_identity.public_key_bytes(),
-            destination_identity.verifying_key_bytes(),
-        );
-        let uri = wire
-            .pack_paper_uri_with_rng(&destination_identity, OsRng)
-            .map_err(std::io::Error::other)?;
-        Ok(Some(PaperEncodeEnvelope {
-            uri,
-            transient_id,
-            destination_hint: record.destination.clone(),
-            extensions: serde_json::Map::new(),
-        }))
+        paper::encode_paper(self, record)
     }
 
     fn decode_paper_uri(&self, uri: &str) -> Result<Option<PaperDecodeOutcome>, std::io::Error> {
-        let wire =
-            WireMessage::unpack_paper_uri(uri, &self.signer).map_err(std::io::Error::other)?;
-        let raw_lxmf_bytes = wire.pack().map_err(std::io::Error::other)?;
-        let transient_id = hex::encode(wire.message_id());
-        let destination_hint = hex::encode(wire.destination);
-        let record = rns_rpc::MessageRecord {
-            id: transient_id.clone(),
-            source: hex::encode(wire.source),
-            destination: destination_hint.clone(),
-            title: wire
-                .payload
-                .title
-                .as_ref()
-                .map(|title| String::from_utf8_lossy(title).to_string())
-                .unwrap_or_default(),
-            content: wire
-                .payload
-                .content
-                .as_ref()
-                .map(|content| String::from_utf8_lossy(content).to_string())
-                .unwrap_or_default(),
-            timestamp: wire.payload.timestamp as i64,
-            direction: "in".to_string(),
-            fields: wire.payload.fields.as_ref().and_then(rmpv_to_json),
-            receipt_status: None,
-        };
-        Ok(Some(PaperDecodeOutcome {
-            transient_id,
-            destination_hint,
-            record: Some(record),
-            raw_lxmf_bytes: Some(raw_lxmf_bytes),
-        }))
+        paper::decode_paper_uri(self, uri)
     }
 
     fn deliver(
