+++ crates/apps/reticulumd/src/bin/reticulumd/bridge_remote_control.rs
@@ -1,4 +1,5 @@
 use super::*;
+use reticulum_daemon::lxmf_bridge::rmpv_to_json;
 
 impl TransportBridge {
     pub(super) fn run_remote_control(
